refactor: add count method to Connection trait
- Add count() method to Connection trait
- Add connection_count() helper for default implementation
- Implement count() for all 6 backends (postgres, mysql, oracle, sqlite, dm, mdb)
- Update table.rs fetch_row_count() to use conn.count()
