+++ remote-table/src/connection/mdb/mod.rs
@@ -343,6 +343,17 @@ impl Connection for MdbConnection {
             "Insert operation is not supported for mdb".to_string(),
         ))
     }
+
+    async fn count(
+        &self,
+        conn_options: &ConnectionOptions,
+        source: &RemoteSource,
+        unparsed_filters: &[String],
+        limit: Option<usize>,
+    ) -> DFResult<Option<usize>> {
+        crate::connection::connection_count(self, conn_options, source, unparsed_filters, limit)
+            .await
+    }
 }
 
 /// List user tables in an MDB file using the ODBC `SQLTables` function.
