+++ remote-table/src/connection/sqlite.rs
@@ -194,6 +194,17 @@ impl Connection for SqliteConnection {
 
         Ok(count)
     }
+
+    async fn count(
+        &self,
+        conn_options: &ConnectionOptions,
+        source: &RemoteSource,
+        unparsed_filters: &[String],
+        limit: Option<usize>,
+    ) -> DFResult<Option<usize>> {
+        crate::connection::connection_count(self, conn_options, source, unparsed_filters, limit)
+            .await
+    }
 }
 
 #[derive(Debug)]
