+++ remote-table/src/table.rs
@@ -487,22 +487,7 @@ pub(crate) async fn fetch_row_count(
     unparsed_filters: &[String],
     limit: Option<usize>,
 ) -> DFResult<Option<usize>> {
-    let db_type = conn_options.db_type();
-    let count1_query = if unparsed_filters.is_empty() && limit.is_none() {
-        db_type.try_count1_query(source)
-    } else {
-        let real_sql = db_type.rewrite_query(source, unparsed_filters, limit);
-        db_type.try_count1_query(&RemoteSource::Query(real_sql))
-    };
-
-    if let Some(count1_query) = count1_query {
-        debug!("[remote-table] fetching row count with query: {count1_query}");
-        let conn = pool.get().await?;
-        let row_count = db_type
-            .fetch_count(conn, conn_options, &count1_query)
-            .await?;
-        Ok(Some(row_count))
-    } else {
-        Ok(None)
-    }
+    let conn = pool.get().await?;
+    conn.count(conn_options, source, unparsed_filters, limit)
+        .await
 }
